Renamed LibraryPartyInfo -> LibraryGroup and removed id property. LibraryGroup is now passed as property to Library which is passed as a navigation argument.
