+++ androidutilities/src/main/java/com/heyzeusv/androidutilities/compose/about/AboutScreens.kt
@@ -1,11 +1,12 @@
 package com.heyzeusv.androidutilities.compose.about
 
+import com.heyzeusv.androidutilities.compose.about.library.LibraryGroup
 import kotlinx.serialization.Serializable
 
-sealed interface AboutScreens {
+internal sealed interface AboutScreens {
     @Serializable
     data object Overview : AboutScreens
 
     @Serializable
-    data class Library(val partyId: String, val libraryId: String) : AboutScreens
+    data class Library(val libraryGroup: LibraryGroup, val libraryId: String) : AboutScreens
 }
