+++ androidutilities/src/main/java/com/heyzeusv/androidutilities/compose/about/LibraryListState.kt
@@ -1,9 +1,10 @@
-package com.heyzeusv.androidutilities.compose.about.library
+package com.heyzeusv.androidutilities.compose.about
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.State
 import androidx.compose.runtime.produceState
 import androidx.compose.ui.platform.LocalContext
+import com.heyzeusv.androidutilities.compose.about.library.LibraryGroup
 import com.mikepenz.aboutlibraries.Libs
 import com.mikepenz.aboutlibraries.entity.Library
 import com.mikepenz.aboutlibraries.util.withContext
@@ -16,19 +17,19 @@ private const val GOOGLE = "com.google"
 private val firstPartyIds = listOf(ANDROID, JETBRAINS, GOOGLE)
 
 @Composable
-fun produceLibraryState(separateByParty: Boolean): State<Map<LibraryPartyInfo, List<Library>>> {
+fun produceLibraryListState(separateByParty: Boolean): State<Map<LibraryGroup, List<Library>>> {
     val context = LocalContext.current
 
-    return produceState(mapOf(LibraryPartyInfo.INIT to listOf())) {
+    return produceState(mapOf(LibraryGroup.INIT to listOf())) {
         value = withContext(Dispatchers.IO) {
             val libs = Libs.Builder().withContext(context).build()
             if (separateByParty) {
                 val (thirdLibs, firstLibs) = libs.libraries.partition { library ->
                     !firstPartyIds.any { library.uniqueId.contains(it) }
                 }
-                mapOf(LibraryPartyInfo.THIRD to thirdLibs, LibraryPartyInfo.FIRST to firstLibs)
+                mapOf(LibraryGroup.THIRD to thirdLibs, LibraryGroup.FIRST to firstLibs)
             } else {
-                mapOf(LibraryPartyInfo.ALL to libs.libraries)
+                mapOf(LibraryGroup.ALL to libs.libraries)
             }
         }
     }
